Challenge3/requirement.txt
ucimlrepo 
fastapi
uvicorn 
aiofiles 
python-multipart 
jinja2
pydantic
pydantic-core
pyserial